requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aioscheduler==1.4.2
aiosignal==1.3.1
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
psycopg2==2.9.9
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.28
